requirements.txt
# FastAPI
fastapi>=0.109.0
uvicorn[standard]>=0.27.0

# MySQL
PyMySQL>=1.1.0
mysql-connector-python>=8.2.0

# Meilisearch
meilisearch>=0.31.0

# Config & validation
pydantic>=2.5.0
pydantic-settings>=2.1.0

# HTTP client (health checks)
httpx>=0.26.0

# AWS upload utilities
boto3>=1.34.0
